Use the `sourcemeta_enable_simd()` in Blaze
I'm making SIMD flags opt-in to make it easier to release cross-platform
binaries for tools like the JSON Schema CLI (as some target machines
will not support SIMD).

While we don't use SIMD directly, auto-vectorisation does, and saw
issues in the wild with the JSON Schema CLI binaries not working on
certain machines.

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>