Add benchmarking with JSON Model Compiler and JSON Schema Utils
Four implementations are added based on [JSON Model Compiler](https://github.com/clairey-zx81/json-model):

- jmc C (jansson, re2 or pcre2)
- jmc JavaScript
- jmc Python (re2 or default)
- jmc Perl (re2 or default)

The benchmark script uses JSU tools to simplify the input schemas and convert them to models,
before compiling them and running on the test values.

The Perl proof-of-concept implementation is ignored as it is very slow.