Avoid top-level `$ref` in Draft 7 and older
As per the specification, a subschema defining `$ref` is a "ref object"
and EVERY other property MUST be completely ignored.

Many implementations (including Blaze), tried to be a bit lenient in
these cases, mainly at the root of a schema.

However, after discussing some pathological cases, the TSC came into
agreement that ANY attempt to make this work, even at the root, can lead
to undefined behavior. Therefore, many implementations, including Blaze,
will now be more strict on this, completely rejecting Draft 7 and older
schemas with a top-level `$ref`.

In this PR, I update a few of the test cases that hit this problem. I
modified them to make them work exactly like before, so we are not
testing more or less than we did before given the workarounds we had in
place.

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>